Extract testfile for tests
