+++ LCGTests/LCGTests.csproj
@@ -229,8 +229,9 @@
   <ItemGroup>
     <Compile Include="CSharpUtilTests.cs" />
     <Compile Include="DatePreservationTests.cs" />
-    <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="LCGHelperTest.cs" />
+    <Compile Include="TestContentHelper.cs" />
+    <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
@@ -240,6 +241,9 @@
     <None Include="testdata\LateboundConstantsGeneratorConfiguration.xml">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
+    <Content Include="testdata\ExpectedTestContent.cs">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\LCGCmd\LCGCmd.csproj">
