+++ LCGTests/testdata/ExpectedTestContent.cs
@@ -1,0 +1,16 @@
+// *********************************************************************
+// Created by : Latebound Constants Generator 1.2023.5.901 for XrmToolBox
+// Tool Author: Jonas Rapp https://jonasr.app/
+// GitHub     : https://github.com/rappen/LCG-UDG/
+// Source Org : https://test.crm.dynamics.com
+// Filename   : {0}
+// Created    : {1}
+// *********************************************************************
+
+namespace TestNamespace
+{{
+    public class TestClass
+    {{
+        public const string EntityName = ""test_entity"";
+    }}
+}}
